+++ ReactWindows/ReactNative.Shared.Tests/ReactNative.Shared.Tests.projitems
@@ -9,7 +9,6 @@
     <Import_RootNamespace>ReactNative.Shared.Tests</Import_RootNamespace>
   </PropertyGroup>
   <ItemGroup>
-    <Compile Include="$(MSBuildThisFileDirectory)Bridge\JavaScriptModuleBaseTests.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Bridge\JavaScriptModuleRegistryTests.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Bridge\ReactContextNativeModuleBaseTests.cs" />
   </ItemGroup>
