+++ ReactWindows/ReactNative.Tests/ReactNative.Tests.csproj
@@ -180,6 +180,7 @@
       <Name>ReactNative</Name>
     </ProjectReference>
   </ItemGroup>
+  <Import Project="..\ReactNative.Shared.Tests\ReactNative.Shared.Tests.projitems" Label="Shared" />
   <PropertyGroup Condition=" '$(VisualStudioVersion)' == '' or '$(VisualStudioVersion)' &lt; '14.0' ">
     <VisualStudioVersion>14.0</VisualStudioVersion>
   </PropertyGroup>
